Add acceptance test
