+++ acceptance/openstack/smn/v2/topicattributes_test.go
@@ -4,11 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/clients"
-	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/tools"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/smn/v2/topicattributes"
-	"github.com/opentelekomcloud/gophertelekomcloud/openstack/smn/v2/topics"
 	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 )
 
@@ -32,20 +29,6 @@ func TestTopicAttributeWorkflow(t *testing.T) {
 	th.AssertEquals(t, examplePolicy, attributes[attribute])
 
 	err = topicattributes.Delete(client, topic, attribute).ExtractErr()
-	th.AssertNoErr(t, err)
-}
-
-func createTopic(t *testing.T, client *golangsdk.ServiceClient) string {
-	opts := topics.CreateOps{
-		Name: tools.RandomString("topic-", 3),
-	}
-	topic, err := topics.Create(client, opts).Extract()
-	th.AssertNoErr(t, err)
-	return topic.TopicUrn
-}
-
-func deleteTopic(t *testing.T, client *golangsdk.ServiceClient, topicURN string) {
-	err := topics.Delete(client, topicURN).ExtractErr()
 	th.AssertNoErr(t, err)
 }
